fix: Empty query are now more likely to display no results.
I delay the last query update (the empty query) with settimeout.
Thus to let a chance to all previous query request to finish.
I tried to use debounce but I was not happy with the result,
the UX felt slow and I felt I had a need to use a loader, I
prefered the input field to be fast. Definitely an issue to look
at later.

Otherwise I added the autofocus on the input field as it was a bit
annoying to have to click on it all the time.